rename: Move Sidebar files under new folder at components/Sidebar
